Fix iptables bug

+++ Back/src/main/java/com/aberdote/OVPN4ALL/service/impl/ConfigServiceImpl.java
@@ -4,6 +4,7 @@ package com.aberdote.OVPN4ALL.service.impl;
 import com.aberdote.OVPN4ALL.dto.SetupDTO;
 import com.aberdote.OVPN4ALL.exception.CustomException;
 import com.aberdote.OVPN4ALL.repository.ConfigRepository;
+import com.aberdote.OVPN4ALL.service.CommandService;
 import com.aberdote.OVPN4ALL.service.ConfigService;
 import com.aberdote.OVPN4ALL.utils.validator.config.ConfigValidator;
 import com.aberdote.OVPN4ALL.utils.validator.converter.EntityConverter;
@@ -13,11 +14,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.io.IOException;
 
 @Slf4j @Service @Transactional @RequiredArgsConstructor
 public class ConfigServiceImpl implements ConfigService {
 
     private final ConfigRepository configRepository;
+    private final CommandService commandService;
 
     @Override
     public SetupDTO getConfig() {
@@ -47,10 +50,20 @@ public class ConfigServiceImpl implements ConfigService {
             log.error("Cannot save setup, netmask is not valid");
             throw new CustomException("Netmask is not valid", HttpStatus.BAD_REQUEST);
         }
-        if (!configRepository.findAll().isEmpty()) configRepository.deleteAll();
-        log.info("Saving new setup");
-        configRepository.save(EntityConverter.fromSetupDTOToConfigEntity(setupDTO));
-        return setupDTO;
+        try {
+            if (!commandService.addConfig(setupDTO.getPort(), setupDTO.getGateway(), setupDTO.getSubnet())) {
+                throw new CustomException("Cannot setup config", HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            log.info("Setting config");
+            if (!configRepository.findAll().isEmpty()) configRepository.deleteAll();
+            log.info("Saving new setup");
+            configRepository.save(EntityConverter.fromSetupDTOToConfigEntity(setupDTO));
+            return setupDTO;
+        } catch (IOException | InterruptedException e) {
+            final String message = String.format("Cannot execute set config script, ErrorMessage: '%s'", e.getMessage());
+            log.error(message);
+            throw new CustomException(message, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
 }
